Backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic[email]
python-dotenv==1.0.0
psycopg2-binary==2.9.9
email-validator==2.1.0